Modified to compile with latest WPI Library.

+++ Source/2009/CompetitionRobot/Potentiometer.cpp
@@ -44,7 +44,7 @@ Potentiometer::Potentiometer(UINT32 channel, float min, float max)
 }
 
 
-float Potentiometer::PIDGet(void)
+double Potentiometer::PIDGet(void)
 {
   return GetAverageVoltage();
 }
